requirements.txt
streamlit==1.38.0
facenet-pytorch==2.6.0
opencv-python==4.10.0.84
numpy>=1.26.0  # Use newer version compatible with Python 3.13
pillow==10.4.0
scipy>=1.13.0  # Use newer version compatible with Python 3.13
scikit-learn>=1.5.0  # Use newer version compatible with Python 3.13
joblib>=1.4.0
faster-whisper==1.0.3
TTS==0.22.0
ollama==0.3.3
webrtcvad==2.0.10
pyaudio==0.2.14
flask==2.3.3
flask-socketio==5.3.6
python-socketio==5.10.0
pyinstaller==6.3.0
pywin32==306  # For Windows installer
cryptography>=41.0.0
qrcode>=7.4.2
requests>=2.31.0
sv-ttk>=2.6.0
pyttsx3>=2.90
